Fix type errors in utility functions
- Change getTypebytes return type from float to size_t
- Change getTensorbytes return type from int to size_t
- Fix INT4 type byte count from 0.5 to 1
- Add default return for unknown types in getTypebytes

+++ TRTInfer/utility.h
@@ -27,14 +27,14 @@ namespace utility
      *@param type The basic types of  tensorrt
      *@return byte count
      */
-    TRTInfer_API float getTypebytes(const nvinfer1::DataType &type);
+    TRTInfer_API size_t getTypebytes(const nvinfer1::DataType &type);
 
     /**
      *@brief Input the dimensions and basic types of a matrix, and output the total number of bytes in this matrix
      *@param dim Dimension of  dim matrix
      *@param type The basic data types of  TensorRT
      */
-    TRTInfer_API int getTensorbytes(const nvinfer1::Dims &dim, const nvinfer1::DataType &type);
+    TRTInfer_API size_t getTensorbytes(const nvinfer1::Dims &dim, const nvinfer1::DataType &type);
 
     /**
      * @brief convert the cv type to nvidia tensorrt type
